Writing tests for cloned checklist admin and personal checklist

See merge request c1989618/ons-client-project-group-8!9

+++ src/main/java/ons/group8/repositories/RoleRepositoryJPA.java
@@ -13,5 +13,5 @@ public interface RoleRepositoryJPA extends JpaRepository<Role, Long> {
 
     List<Role> findAll();
 
-    Optional<Role> findById(int id);
+    Optional<Role> findById(long id);
 }
